Added error handling in fetch
